Fixed the inplace bug

+++ Signal/General/dsp/TScrunch.h
@@ -15,7 +15,7 @@ namespace dsp {
 
     TScrunch(Behaviour _type=Operation::anyplace);// : Operation ("TScrunch", _type);
     
-    void set_ScrunchFactor( int _ScrunchFactor){ ScrunchFactor = _ScrunchFactor; use_tres = false; }
+    void set_ScrunchFactor( int64 _ScrunchFactor){ ScrunchFactor = _ScrunchFactor; use_tres = false; }
     int64 get_ScrunchFactor(){ return ScrunchFactor; }
 
     void set_NewTimeRes( double microseconds );
